Add docs for datasets and hide some private methods
