+++ faculty/datasets/util.py
@@ -12,36 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import posixpath
+"""Common components for Faculty datasets."""
 
 
 class DatasetsError(Exception):
+    """An error occurred when using Faculty datasets."""
+
     pass
-
-
-def rationalise_path(path):
-
-    # All paths should be relative to root
-    path = posixpath.join("/", path)
-
-    normed = posixpath.normpath(path)
-
-    if path.endswith("/") and not normed.endswith("/"):
-        normed += "/"
-
-    return normed
-
-
-def get_relative_path(parent_directory, directory):
-
-    parent_directory = rationalise_path(parent_directory)
-    directory = rationalise_path(directory)
-
-    if not directory.startswith(parent_directory):
-        tpl = "{} is not a sub path of {}"
-        raise ValueError(tpl.format(directory, parent_directory))
-
-    # Remove the root
-    relative_path = posixpath.relpath(directory, parent_directory)
-
-    return relative_path
