--- tests/datasets/test_util.py
@@ -1,35 +1,0 @@
-# Copyright 2018-2020 Faculty Science Limited
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import pytest
-
-from faculty.datasets import util
-
-
-@pytest.mark.parametrize(
-    "input_path, rationalised_path",
-    [
-        ("", "/"),
-        ("./", "/"),
-        ("/", "/"),
-        ("path", "/path"),
-        ("./path", "/path"),
-        ("/path", "/path"),
-        ("path/", "/path/"),
-        ("./path/", "/path/"),
-        ("/path/", "/path/"),
-    ],
-)
-def test_rationalise_path(input_path, rationalised_path):
-    assert util.rationalise_path(input_path) == rationalised_path
